requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
gunicorn==23.0.0
celery[redis]==5.4.0
redis==5.2.1
elasticsearch==8.17.0
SQLAlchemy==2.0.37
pydantic==2.10.5
pydantic-settings==2.7.1
openai==1.60.1
httpx==0.28.1
watchdog==6.0.0
pdfplumber==0.11.5
python-docx==1.1.2
orjson==3.10.13
python-dotenv==1.0.1
pytest==8.3.4
